Address a ton of circular dependencies
