+++ packages/codec/lib/encode/key.ts
@@ -1,8 +1,9 @@
 import debugModule from "debug";
 const debug = debugModule("codec:encode:key");
 
-import { Values } from "../format";
-import { Conversion as ConversionUtils, EVM as EVMUtils } from "../utils";
+import { Values } from "../format/values";
+import { Conversion as ConversionUtils } from "../utils/conversion";
+import { EVM as EVMUtils } from "../utils/evm";
 import { stringToBytes } from "./abi";
 import BN from "bn.js";
 
@@ -57,71 +58,4 @@ export function encodeMappingKey(input: Values.ElementaryValue): Uint8Array {
 
 export function mappingKeyAsHex(input: Values.ElementaryValue): string {
   return ConversionUtils.toHexString(encodeMappingKey(input));
-}
-
-//this is like the old toSoliditySha3Input, but for debugging purposes ONLY
-//it will NOT produce correct input to soliditySha3
-//please use mappingKeyAsHex instead if you wish to encode a mapping key.
-export function keyInfoForPrinting(input: Values.ElementaryValue): {type: string, value: string} {
-  switch(input.type.typeClass) {
-    case "uint":
-      return {
-        type: "uint",
-        value: (<Values.UintValue>input).value.asBN.toString()
-      };
-    case "int":
-      return {
-        type: "int",
-        value: (<Values.IntValue>input).value.asBN.toString()
-      };
-    case "fixed":
-      return {
-        type: `fixed256x${input.type.places}`,
-        value: (<Values.FixedValue>input).value.asBig.toString()
-      };
-    case "ufixed":
-      return {
-        type: `ufixed256x${input.type.places}`,
-        value: (<Values.UfixedValue>input).value.asBig.toString()
-      };
-    case "bool":
-      //this is the case that won't work as valid input to soliditySha3 :)
-      return {
-        type: "uint",
-        value: (<Values.BoolValue>input).value.asBool.toString()
-      };
-    case "bytes":
-      switch(input.type.kind) {
-        case "static":
-          return {
-            type: "bytes32",
-            value: (<Values.BytesValue>input).value.asHex
-          };
-        case "dynamic":
-          return {
-            type: "bytes",
-            value: (<Values.BytesValue>input).value.asHex
-          };
-      }
-    case "address":
-      return {
-        type: "address",
-        value: (<Values.AddressValue>input).value.asAddress
-      };
-    case "string":
-      let coercedInput: Values.StringValue = <Values.StringValue> input;
-      switch(coercedInput.value.kind) {
-        case "valid":
-          return {
-            type: "string",
-            value: coercedInput.value.asString
-          };
-        case "malformed":
-          return {
-            type: "bytes",
-            value: coercedInput.value.asHex
-          };
-      }
-    //fixed and ufixed are skipped for now
-  }
 }
