+++ packages/codec/lib/read/stack.ts
@@ -2,11 +2,11 @@ import debugModule from "debug";
 const debug = debugModule("codec:read:stack");
 
 import * as CodecUtils from "../utils";
-import { Errors } from "../format/errors";
+import { DecodingError } from "../types/errors";
 
 export function readStack(stack: Uint8Array[], from: number, to: number): Uint8Array {
   if(from < 0 || to >= stack.length) {
-    throw new Errors.DecodingError(
+    throw new DecodingError(
       {
         kind: "ReadErrorStack",
         from,
