+++ packages/codec/lib/types/decoding.ts
@@ -1,6 +1,7 @@
 import { DecoderContext } from "./contexts";
 import * as AbiTypes from "./abi";
-import { Types, Values } from "../format";
+import { Types } from "../format/types";
+import { Values } from "../format/values";
 
 export type CalldataDecoding = FunctionDecoding | ConstructorDecoding | MessageDecoding | UnknownDecoding;
 export type LogDecoding = EventDecoding | AnonymousDecoding;
