+++ packages/codec/lib/types/errors.ts
@@ -1,4 +1,5 @@
 import { Errors } from "../format";
+import { message } from "../utils/errors";
 
 export class UnknownBaseContractIdError extends Error {
   public derivedId: number;
@@ -13,6 +14,18 @@ export class UnknownBaseContractIdError extends Error {
     this.derivedName = derivedName;
     this.derivedKind = derivedKind;
     this.baseId = baseId;
+  }
+}
+
+//For when we need to throw an error, here's a wrapper class that extends Error.
+//Apologies about the confusing name, but I wanted something that would make
+//sense should it not be caught and thus accidentally exposed to the outside.
+export class DecodingError extends Error {
+  public error: Errors.ErrorForThrowing;
+  constructor(error: Errors.ErrorForThrowing) {
+    super(message(error));
+    this.error = error;
+    this.name = "DecodingError";
   }
 }
